requirements.txt
python-binance
pandas
pandas_ta
python-dotenv